changed from integer to double, percentage can be float

+++ src/main/java/com/gwandubruce/dronesapi/models/BatteryLevelEventLog.java
@@ -23,7 +23,7 @@ public class BatteryLevelEventLog {
     @ManyToOne
     private Drone drone;
 
-    private Integer batteryLevel;
+    private Double batteryLevel;
 
     private LocalDateTime timeOfLogEntry;
 }
